added a better routes
