+++ src/admin/routes/AdminRoutes.js
@@ -1,12 +1,15 @@
 import React from 'react'
 import { BrowserRouter, Route } from 'react-router-dom'
-import Sidebar from '../component/Sidebar'
+import AppWrapper from '../component/AppWrapper'
 
 
 export default function AdminRoutes(){
     return(
         <BrowserRouter>
-            <Route path='/admin/overview' component={Sidebar} />
+            <AppWrapper>
+                <Route exact path='/admin/overview' component={() => <h4>Overview</h4>} />
+                <Route exact path='/admin/users' component={() => <h4>Users</h4>} />
+            </AppWrapper>
         </BrowserRouter>
     )
 }
