+++ src/admin/styles/styles.css
@@ -25,6 +25,7 @@
   margin-left: 160px; /* Same as the width of the sidenav */
   font-size: 28px; /* Increased text to enable scrolling */
   padding: 0px 10px;
+  margin-top: 6%;
 }
 
 @media screen and (max-height: 450px) {
