+++ src/core/routes/Routes.js
@@ -1,11 +1,15 @@
 import React from 'react'
+import { BrowserRouter, Route } from 'react-router-dom'
 import Register from '../component/Auth/Register'
-import SimpleSidebar from '../../admin/component/Sidebar'
+// import SimpleSidebar from '../../admin/component/Sidebar'
+import AdminRoutes from '../../admin/routes/AdminRoutes'
 
 function Routes(){
     return(
-        // <Register />
-        <SimpleSidebar />
+        <BrowserRouter>
+            <Route  path='/admin/' component={AdminRoutes} />
+            <Route exact path='/register' component={Register} />
+        </BrowserRouter>
     )
 }
 export default Routes
